requirements.txt
langchain>=0.0.8
chromadb>=0.3.22
memgpt
langgraph
langmem
openai
pytest
loguru
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
letta
matplotlib